% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{.calibrate_eta_gpc}
\alias{.calibrate_eta_gpc}
\title{Internal wrapper for GPC calibration}
\usage{
.calibrate_eta_gpc(
  x,
  side,
  target,
  content,
  confidence,
  tau,
  tau_lower,
  tau_upper,
  verbose,
  eta0,
  max_iter,
  B,
  tol,
  c,
  gamma,
  prop_sd,
  w,
  m,
  n_samps_boot,
  burn_in_boot
)
}
\description{
Internal wrapper for GPC calibration
}
