% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{.run_sampler}
\alias{.run_sampler}
\title{Internal wrapper for final MCMC sampling}
\usage{
.run_sampler(
  x,
  side,
  target,
  tau,
  tau_lower,
  tau_upper,
  eta,
  n_mcmc,
  burnin,
  verbose
)
}
\description{
Internal wrapper for final MCMC sampling
}
